+++ package.json
@@ -43,7 +43,7 @@
         "@gitbeaker/rest": "^42.5.0",
         "@golevelup/nestjs-rabbitmq": "^6.0.1",
         "@google/generative-ai": "^0.24.1",
-        "@kodus/kodus-proto": "^2.0.3",
+        "@kodus/kodus-proto": "2.0.9",
         "@langchain/anthropic": "^0.3.21",
         "@langchain/cohere": "^0.3.3",
         "@langchain/community": "^0.3.45",
@@ -125,7 +125,7 @@
         "nestjs-pino": "^4.4.0",
         "nodemailer-express-handlebars": "^7.0.0",
         "octokit": "^5.0.3",
-        "openai": "^5.1.1",
+        "openai": "^5.2.0",
         "opossum": "^9.0.0",
         "p-limit": "^6.2.0",
         "passport": "^0.7.0",
@@ -172,8 +172,8 @@
         "@types/passport-jwt": "^4.0.1",
         "@types/ramda": "^0.30.2",
         "@types/supertest": "^6.0.3",
-        "@typescript-eslint/eslint-plugin": "^8.33.1",
-        "@typescript-eslint/parser": "^8.33.1",
+        "@typescript-eslint/eslint-plugin": "^8.34.0",
+        "@typescript-eslint/parser": "^8.34.0",
         "concurrently": "^9.1.2",
         "esbuild": "^0.25.5",
         "eslint": "^9.28.0",
