+++ src/core/domain/codeBase/contracts/ASTAnalysisService.contract.ts
@@ -5,12 +5,12 @@ import {
     ReviewModeResponse,
 } from '@/config/types/general/codeReview.type';
 import { OrganizationAndTeamData } from '@/config/types/general/organizationAndTeamData';
+import { ChangeResult } from '@/ee/codeBase/types/diff-analyzer.types';
 
 import {
     FunctionsAffectResult,
     FunctionSimilarity,
 } from '@/ee/kodyAST/code-analyzer.service';
-import { ChangeResult } from '@/ee/codeBase/diffAnalyzer.service';
 
 export const AST_ANALYSIS_SERVICE_TOKEN = Symbol('ASTAnalysisService');
