+++ src/modules/codeReviewPipeline.module.ts
@@ -18,7 +18,6 @@ import { ParametersModule } from './parameters.module';
 import { PlatformIntegrationModule } from './platformIntegration.module';
 import { AutomationModule } from './automation.module';
 import { PullRequestsModule } from './pullRequests.module';
-import { KodyASTModule } from '../ee/kodyAST/kodyAST.module';
 import { KodyRulesModule } from './kodyRules.module';
 import { SuggestionEmbeddedModule } from './suggestionEmbedded.module';
 import { OrganizationParametersModule } from './organizationParameters.module';
@@ -44,7 +43,6 @@ import { KodyASTAnalyzeContextModule } from '@/ee/kodyASTAnalyze/kodyAstAnalyzeC
         forwardRef(() => FileReviewModule),
         forwardRef(() => KodyFineTuningContextModule),
         forwardRef(() => KodyASTAnalyzeContextModule),
-        KodyASTModule.register(),
     ],
     providers: [
         PipelineExecutor,
